+++ ezdxf/__init__.py
@@ -18,7 +18,7 @@ from ezdxf.lldxf import const  # restore module structure ezdxf.const
 from ezdxf.lldxf.validator import is_dxf_file, is_dxf_stream
 from ezdxf.filemanagement import new, read, readfile, readzip
 from ezdxf.tools.standards import setup_linetypes, setup_styles, setup_dimstyles, setup_dimstyle
-from ezdxf.lldxf.const import Arrows
+from ezdxf.render.arrows import ARROWS
 
 # Exceptions
 from ezdxf.lldxf.const import DXFError  # base error exception