+++ ezdxf/addons/__init__.py
@@ -3,11 +3,6 @@
 # License: MIT License
 from .table import Table, CustomCell
 from .mtext import MText
-from .curves import Bezier, EulerSpiral, Spline
-from .r12spline import R12Spline
-from .forms import circle, ellipse, cylinder, cone
-from .forms import extrude, from_profiles_linear, from_profiles_spline, rotation_form
-from .mesh import MeshBuilder, MeshVertexMerger
 from .menger_sponge import MengerSponge
 from .sierpinski_pyramid import SierpinskyPyramid
 from .dimlines import LinearDimension, AngularDimension, ArcDimension, RadialDimension, dimstyles