+++ ezdxf/modern/tableentries.py
@@ -10,7 +10,7 @@ from ezdxf.lldxf.const import DXFKeyError
 from ezdxf.legacy import tableentries as legacy
 from ezdxf.dxfentity import DXFEntity
 from ezdxf.tools.complex_ltype import lin_compiler
-from ezdxf.lldxf.const import Arrows
+from ezdxf.render.arrows import ARROWS
 
 logger = logging.getLogger('ezdxf')
 
@@ -393,7 +393,7 @@ class DimStyle(legacy.DimStyle):
     def set_blocks(self, blk: str = '', blk1: str = '', blk2: str = '') -> None:
         def set_blk_handle(attr: str) -> None:
             blk_name = self.get_dxf_attrib(attr)
-            if blk_name in Arrows:  # not real blocks
+            if ARROWS.is_acad_arrow(blk_name):  # not real blocks
                 return
             blk = blocks.get(blk_name)
             self.set_dxf_attrib(attr+'_handle', blk.block_record_handle)